+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Remove failure reason `connection_refused` and report it as `connection_error` along with the rest of socket related errors.
 - Remove `dnsexp_failures_total` from per-scrape metrics. Failure reasons are now tracked only under `/metrics`.
 - Switch from `http.server.HTTPServer` to `http.server.ThreadingHTTPServer`
+- Add new failure reason `invalid_response_statuscode` for DoH failures.
 
 
 ## [v1.0.0-beta6] - 2024-03-01